Add requiresGrad and default that to false.
It will be moved to true when added to an optimizer. Without
requiresGrad, the variable.grad won't be filled and the computation
won't be executed to save the computation time. It still maintains the
computation graph. For things that doesn't maintain the computation
graph, you need to mark them as constant from the beginning (using
graph.constant).